Sucurl By empresa

+++ src/controllers/sucursalController.ts
@@ -1,3 +1,4 @@
+import { Empresa } from "../models/empresaModel";
 import { Sucursal } from "../models/sucursalModel";
 import { Request, Response } from "express";
 
@@ -36,7 +37,7 @@ export const getById = async (req: Request, res: Response) => {
   try {
     const sucursal = await Sucursal.findById(sucursalId);
     if (!sucursal) {
-      throw "La sucursal no existe";
+      return res.status(400).json({ error: "Sucursal no existe" });
     }
     return res.status(200).json({
       message: "Sucursal obtenida",
@@ -79,9 +80,15 @@ export const update = async (req: Request, res: Response) => {
   const newInfosucursal = req.body;
   try {
     const sucursalValidador = await Sucursal.find({
-      $or: [{ sucursal: newInfosucursal.sucursal }, { ubicacion: newInfosucursal.ubicacion }],
-    })
-    if(sucursalValidador.length && sucursalValidador.some((item)=> item._id != sucursalId)){
+      $or: [
+        { sucursal: newInfosucursal.sucursal },
+        { ubicacion: newInfosucursal.ubicacion },
+      ],
+    });
+    if (
+      sucursalValidador.length &&
+      sucursalValidador.some((item) => item._id != sucursalId)
+    ) {
       return res
         .status(400)
         .json({ error: "Ya existe una empresa con esas caracteristicas" });
@@ -106,13 +113,37 @@ export const update = async (req: Request, res: Response) => {
 export const eliminate = async (req: Request, res: Response) => {
   const sucursalId = req.params.sucursal;
   try {
-    const updatedSucursal = await Sucursal.findByIdAndDelete(sucursalId);
+    const sucursal = await Sucursal.findOne({ sucursal: sucursalId });
+    if (!sucursal) {
+      return res.status(400).json({ error: "Sucursal no existe" });
+    }
+    const deletedSucursal = await Sucursal.findByIdAndDelete(sucursalId);
     return res.status(200).json({
       message: "Sucursal borrada",
     });
   } catch (e) {
     res.status(500).json({
       message: "Error al eliminar sucursal",
+      content: e.message,
+    });
+  }
+};
+
+export const getAllBySucursal = async (req: Request, res: Response) => {
+  const sucursalId = req.params.sucursal;
+  try {
+    const sucursal = await Sucursal.findOne({ sucursal: sucursalId });
+    if (!sucursal) {
+      return res.status(400).json({ error: "Sucursal no existe" });
+    }
+    const sucursales = await Sucursal.find({ empresa_id: sucursal.empresa_id });
+    return res.status(200).json({
+      message: "Lista de sucursales",
+      content: sucursales,
+    });
+  } catch (e) {
+    res.status(500).json({
+      message: "Error al obtener todas las sucursales",
       content: e.message,
     });
   }
